# docs/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Audiobook Librarian API",
        "version": "1.0.0",
        "description": "API documentation for the Audiobook Librarian application.\n\nThis specification covers all endpoints available in the Audiobook Librarian API v1. Authentication uses opaque Bearer tokens returned by the login endpoint.",
        "contact": {
            "name": "API Support",
            "email": "support@librarian.local"
        }
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Local development server"
        },
        {
            "url": "https://books.thelin.org/api/v1",
            "description": "Production server"
        }
    ],
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "Token",
                "description": "Authentication using Bearer tokens. After logging in, include the token in the Authorization header: `Authorization: Bearer {token}`"
            }
        },
        "schemas": {
...
            "Bookmark": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "Bookmark ID",
                        "example": 1
                    },
                    "book_id": {
                        "type": "integer",
                        "description": "Relationship to book",
                        "example": 123
                    },
                    "position_ms": {
                        "type": "integer",
                        "format": "int64",
                        "description": "Position in the book in milliseconds",
                        "example": 120000
                    },
                    "title": {
                        "type": "string",
                        "nullable": true,
                        "description": "Optional title for the bookmark",
                        "example": "Start of Chapter 2"
                    },
                    "note": {
                        "type": "string",
                        "nullable": true,
                        "description": "Optional note",
                        "example": "Important plot point here"
                    },
                    "is_auto": {
                        "type": "boolean",
                        "description": "Whether this was an automatically created bookmark",
                        "example": false
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "description": "Creation timestamp",
                        "example": "2024-01-01T12:00:00Z"
                    }
                },
                "required": [
                    "id",
                    "book_id",
                    "position_ms",
                    "created_at"
                ]
            },
            "CreateBookmarkRequest": {
                "type": "object",
                "properties": {
                    "position_ms": {
                        "type": "integer",
                        "format": "int64",
                        "description": "Position in the book in milliseconds",
                        "example": 120000
                    },
                    "title": {
                        "type": "string",
                        "nullable": true,
                        "description": "Optional title",
                        "example": "Interesting quote"
                    },
                    "note": {
                        "type": "string",
                        "nullable": true,
                        "description": "Optional note",
                        "example": "I should revisit this"
                    },
                    "is_auto": {
                        "type": "boolean",
                        "default": false,
                        "description": "Whether this is an automatic bookmark"
                    }
                },
                "required": [
                    "position_ms"
                ]
            },
...
            "Error": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string",
                        "description": "Error code for identifying the type of error",
                        "example": "ACCOUNT_PENDING_APPROVAL"
                    },
                    "error": {
                        "type": "string",
                        "description": "Error type",
                        "example": "Unauthorized"
                    },
                    "message": {
                        "type": "string",
                        "description": "Detailed error message",
                        "example": "Invalid credentials"
                    }
                }
            },
            "ValidationError": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "description": "Email validation errors",
                        "example": ["The email field is required."]
                    },
                    "password": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "description": "Password validation errors",
                        "example": [
                            "The password must be at least 8 characters."
                        ]
                    }
                }
            },
            "Book": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "Unique book identifier",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "description": "Book title",
                        "example": "The Fellowship of the Ring"
                    },
                    "author": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "description": "List of authors",
                        "example": ["J.R.R. Tolkien"]
                    },
                    "narrator": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "description": "List of narrators",
                        "example": ["Rob Inglis"]
                    },
                    "series": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "name": {
                                    "type": "string",
                                    "description": "Series name",
                                    "example": "The Lord of the Rings"
                                },
                                "series_number": {
                                    "type": "string",
                                    "nullable": true,
                                    "description": "Book number in series",
                                    "example": "1"
                                }
                            }
                        },
                        "description": "List of series this book belongs to with their position numbers",
                        "example": [
                            {
                                "name": "The Lord of the Rings",
                                "series_number": "1"
                            }
                        ]
                    },
                    "genre": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "description": "List of book genres",
                        "example": ["Fantasy", "Adventure"]
                    },
                    "year": {
                        "type": "integer",
                        "nullable": true,
                        "description": "Publication/release year",
                        "example": 2001
                    },
                    "duration": {
                        "type": "string",
                        "nullable": true,
                        "description": "Total audiobook duration (HH:MM:SS)",
                        "example": "19:39:00"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true,
                        "description": "Book description/summary",
                        "example": "The first volume of The Lord of the Rings..."
                    },
                    "cover_url": {
                        "type": "string",
                        "nullable": true,
                        "description": "URL to cover image",
                        "example": "https://books.thelin.org/api/v1/books/1/cover"
                    },
                    "file_count": {
                        "type": "integer",
                        "description": "Number of audio files",
                        "example": 12
                    },
                    "total_size": {
                        "type": "integer",
                        "description": "Total size in bytes",
                        "example": 687194112
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "description": "Creation timestamp",
                        "example": "2024-01-01T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "description": "Last update timestamp",
                        "example": "2024-01-01T12:00:00Z"
                    },
                    "progress": {
                        "$ref": "#/components/schemas/BookProgress",
                        "nullable": true
                    },
                    "status": {
                        "$ref": "#/components/schemas/UserBookStatus",
                        "nullable": true
                    },
                    "recommendation": {
                        "$ref": "#/components/schemas/UserRecommendation",
                        "nullable": true
                    }
                }
            },
...
            "BookProgress": {
                "type": "object",
                "properties": {
                    "book_id": {
                        "type": "integer",
                        "description": "Book ID"
                    },
                    "device_id": {
                        "type": "string",
                        "description": "Device identifier"
                    },
                    "current_position_seconds": {
                        "type": "integer",
                        "description": "Current position in seconds"
                    },
                    "total_duration_seconds": {
                        "type": "integer",
                        "nullable": true,
                        "description": "Total book duration in seconds"
                    },
                    "progress_percentage": {
                        "type": "number",
                        "format": "float",
                        "description": "Progress percentage (0.00 to 100.00)"
                    },
                    "current_chapter": {
                        "type": "integer",
                        "nullable": true,
                        "description": "Current chapter number"
                    },
                    "current_chapter_name": {
                        "type": "string",
                        "nullable": true,
                        "description": "Current chapter name"
                    },
                    "last_listened_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true,
                        "description": "Last listened timestamp"
                    },
                    "completed": {
                        "type": "boolean",
                        "description": "Whether the book is completed"
                    },
                    "completed_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true,
                        "description": "Completion timestamp"
                    },
                    "formatted_progress": {
                        "type": "string",
                        "description": "Human-readable progress time"
                    },
                    "formatted_duration": {
                        "type": "string",
                        "description": "Human-readable total duration"
                    }
                }
            },
...
        }
    },
    "paths": {
...
        "/bookmarks/{bookId}": {
            "get": {
                "tags": ["Bookmarks"],
                "summary": "Get bookmarks for a book",
                "description": "Returns all bookmarks for the specified book belonging to the authenticated user.",
                "operationId": "getBookmarks",
                "security": [{"BearerAuth": []}],
                "parameters": [
                    {
                        "name": "bookId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "ID of the book"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of bookmarks",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "bookmarks": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Bookmark"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Book not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                }
            },
            "post": {
                "tags": ["Bookmarks"],
                "summary": "Create a bookmark",
                "description": "Creates a new bookmark for the specified book.",
                "operationId": "createBookmark",
                "security": [{"BearerAuth": []}],
                "parameters": [
                    {
                        "name": "bookId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "ID of the book"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateBookmarkRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Bookmark created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Bookmark"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Book not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                }
            }
        },
...
        "/books/{book}": {
            "get": {
                "tags": ["Books"],
                "summary": "Get specific book details",
                "description": "Returns detailed information about a specific audiobook",
                "operationId": "getBook",
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "book",
                        "in": "path",
                        "description": "Book ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Book details retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Book"
                                },
                                "examples": {
                                    "book_details": {
                                        "summary": "Complete book details",
                                        "value": {
                                            "id": 1,
                                            "title": "The Fellowship of the Ring",
                                            "author": ["J.R.R. Tolkien"],
                                            "narrator": ["Rob Inglis"],
                                            "series": "The Lord of the Rings",
                                            "series_number": "1",
                                            "genre": "Fantasy",
                                            "year": 2001,
                                            "duration": "19:39:00",
                                            "description": "The first volume of The Lord of the Rings, The Fellowship of the Ring tells of the fateful power of the One Ring. It begins a magnificent tale of adventure that will plunge the members of the Fellowship of the Ring into a perilous quest and set the stage for the ultimate clash between the powers of good and evil.",
                                            "cover_url": "https://books.thelin.org/api/v1/books/1/cover",
                                            "file_count": 12,
                                            "total_size": 687194112,
                                            "created_at": "2024-01-01T12:00:00Z",
                                            "updated_at": "2024-01-01T12:00:00Z"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Book not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                },
                                "examples": {
                                    "not_found": {
                                        "summary": "Book does not exist",
                                        "value": {
                                            "error": "Book not found"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                },
                                "examples": {
                                    "unauthorized": {
                                        "summary": "Invalid or missing token",
                                        "value": {
                                            "error": "Unauthorized"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}